R9148.fix renderthread

Fixes various rendering bugs on all platforms and makes stuff compile with Qt 5.5.1

This has at least the following issues:
* Post-processing framework doesn't work on OS X due to high DPI
* On Linux, screen.xml "screennumber" does not match X11 screen number
* All wgl/glX extensions were disabled due to glbinding not supporting them

See merge request !16

+++ multitude/Luminous/ProgramGL.cpp
@@ -94,10 +94,10 @@ namespace Luminous
     GLERROR("ShaderGL::compile # glShaderSource");
     glCompileShader(m_handle);
     GLERROR("ShaderGL::compile # glCompileShader");
-    GLint compiled = GL_FALSE;
+    GLint compiled = static_cast<GLint>(GL_FALSE);
     glGetShaderiv(m_handle, GL_COMPILE_STATUS, &compiled);
     GLERROR("ShaderGL::compile # glGetShaderiv");
-    if (compiled != GL_TRUE) {
+    if (compiled != static_cast<GLint>(GL_TRUE)) {
       Radiant::error("Failed to compile shader %s", shader.filename().toUtf8().data());
 
       // Dump info log
@@ -217,7 +217,7 @@ namespace Luminous
     glGetProgramiv(m_handle, GL_LINK_STATUS, &status);
     GLERROR("ProgramGL::link # glGetProgramiv");
 
-    if(status == GL_FALSE) {
+    if(status == static_cast<GLint>(GL_FALSE)) {
       Radiant::error("Failed to link shader program (shaders %s)", program.shaderFilenames().join(", ").toUtf8().data());
       GLsizei len;
       glGetProgramiv(m_handle, GL_INFO_LOG_LENGTH, &len);
